Refactoring code in centreon to be align with centreon-UI
