+++ package.json
@@ -88,7 +88,7 @@
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@centreon/centreon-frontend": "git+https://centreon@github.com/centreon/centreon-frontend.git#develop",
+    "@centreon/centreon-frontend": "git+https://centreon@github.com/centreon/centreon-frontend.git#Align-top-counter-icon-with-centreon-ui-theme",
     "@date-io/dayjs": "1.3.13",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
