+++ www/front_src/src/Header/RessourceStatusCounter/getResourcesUrl.ts
